+++ css-doodle.js
@@ -1,4 +1,4 @@
-/*! css-doodle@0.37.0 */
+/*! css-doodle@0.37.1 */
 (function (global, factory) {
   typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
   typeof define === 'function' && define.amd ? define(factory) :
@@ -952,7 +952,11 @@
       }
     }
     if (headSVG && Array.isArray(headSVG.value)) {
-      headSVG.value.push(...headVariables);
+      for (let variable of headVariables) {
+        if (!headSVG.value.find(n => n.name == variable.name)) {
+          headSVG.value.unshift(variable);
+        }
+      }
       return headSVG;
     }
     return block;
@@ -1476,9 +1480,10 @@
         let [args, raw_args] = read_arguments(it, composition, composible(name), variables);
         if (is_svg(name)) {
           let parsed_svg = parse$7(raw_args);
+          let line = 0;
           for (let item of parsed_svg.value) {
             if (item.variable) {
-              variables[item.name] = (parse$6(`${item.name}: ${item.value}`))[0].value;
+              variables[item.name] = (parse$6(`${'\n'.repeat(line++)} ${item.name}: ${item.value}`))[0].value;
             }
           }
           if (/\d\s*{/.test(raw_args) && has_times_syntax(parsed_svg)) {
@@ -4728,7 +4733,7 @@
       return Expose[name] || MathFunc[name];
     }
 
-    apply_func(fn, coords, args, fname) {
+    apply_func(fn, coords, args, fname, contextedVariable = {}) {
       let _fn = fn(...make_array(coords));
       let input = [];
       args.forEach(arg => {
@@ -4754,7 +4759,7 @@
             this.custom_properties['host'],
             this.custom_properties['container'],
             this.custom_properties[coords.count],
-            coords.context['v-counter'] || {}
+            contextedVariable
           );
           let context = {};
           let unit = '';
@@ -4783,13 +4788,13 @@
       return ['doodle', 'shaders', 'canvas', 'pattern'].includes(name);
     }
 
-    read_var(value, coords) {
+    read_var(value, coords, contextedVariable) {
       let count = coords.count;
       let group = Object.assign({},
         this.custom_properties['host'],
         this.custom_properties['container'],
         this.custom_properties[count],
-        coords.context['v-counter'] || {}
+        contextedVariable
       );
       if (group[value] !== undefined) {
         let result = String(group[value]).trim();
@@ -4804,7 +4809,7 @@
       return value;
     }
 
-    compose_argument(argument, coords, extra = [], parent) {
+    compose_argument(argument, coords, extra = [], parent, contextedVariable) {
       if (!coords.extra) coords.extra = [];
       coords.extra.push(extra);
       let result = argument.map(arg => {
@@ -4813,7 +4818,7 @@
             if (parent && parent.name === '@var') {
               return arg.value;
             }
-            return this.read_var(arg.value, coords);
+            return this.read_var(arg.value, coords, contextedVariable);
           }
           return arg.value;
         }
@@ -4847,11 +4852,10 @@
             coords.position = arg.position;
             let args = arg.arguments.map(n => {
               return fn.lazy
-                ? (...extra) => this.compose_argument(n, coords, extra, arg)
-                : this.compose_argument(n, coords, extra, arg);
+                ? (...extra) => this.compose_argument(n, coords, extra, arg, contextedVariable)
+                : this.compose_argument(n, coords, extra, arg, contextedVariable);
             });
-            let value = this.apply_func(fn, coords, args, fname);
-            return value;
+            return this.apply_func(fn, coords, args, fname, contextedVariable);
           } else {
             return arg.name;
           }
@@ -4930,18 +4934,21 @@
       return value;
     }
 
-    compose_variables(variables, coords) {
-      let counter = 'v-counter';
-      if (!coords.context[counter]) {
-        coords.context[counter] = {};
-      }
+    compose_variables(variables, coords, result = {}) {
       for (let [name, value] of Object.entries(variables)) {
-        coords.context[counter][name] =
-          this.compose_value(value[0], coords, coords.context[counter]).value;
+        let value_group = value.reduce((ret, v) => {
+          let composed = this.compose_value(v, coords, result);
+          if (composed && composed.value) {
+            ret.push(composed.value);
+          }
+          return ret;
+        }, []);
+        result[name] = value_group.join(', ');
       }
+      return result;
     }
 
-    compose_value(value, coords, contextedVariable) {
+    compose_value(value, coords, contextedVariable = {}) {
       if (!Array.isArray(value)) {
         return {
           value: '',
@@ -4984,21 +4991,15 @@
               } else {
                 coords.position = val.position;
                 if (val.variables) {
-                  this.compose_variables(val.variables, coords);
+                  this.compose_variables(val.variables, coords, contextedVariable);
                 }
                 let args = val.arguments.map(arg => {
                   return fn.lazy
-                    ? (...extra) => {
-                        let composed = this.compose_argument(arg, coords, extra, val);
-                        if (val.variables) {
-                          coords.context['v-counter'] = {};
-                        }
-                        return composed;
-                    }
-                    : this.compose_argument(arg, coords, [], val);
+                    ? (...extra) => this.compose_argument(arg, coords, extra, val, contextedVariable)
+                    : this.compose_argument(arg, coords, [], val, contextedVariable);
                 });
 
-                let output = this.apply_func(fn, coords, args, fname);
+                let output = this.apply_func(fn, coords, args, fname, contextedVariable);
                 if (!is_nil(output)) {
                   result += output;
                   if (output.extra) {